svein/flask_barcode/DATASCI210_barcode_scanning/requirements.txt
Flask
flask-cors
Pillow
pyzbar
pandas
tabulate
openpyxl
opencv-python